Refactor authentication logic
