+++ src/tests/api/v1/auth.int.test.ts
@@ -4,7 +4,7 @@ import {
   AppErrorResponse,
   AppJwtPayload,
   NewDefaultUser,
-  SignInResponse,
+  AuthResponse,
 } from '../../../types';
 import bcrypt from 'bcryptjs';
 import jwt from 'jsonwebtoken';
@@ -79,7 +79,7 @@ describe('Authentication endpoint', () => {
 
     it('should sign in and response with JWT and user insensitive-info', async () => {
       const res = await api.post(SIGNIN_URL).send(signInData);
-      const resBody = res.body as SignInResponse;
+      const resBody = res.body as AuthResponse;
       const resUser = resBody.user as User;
       const resJwtPayload = jwt.decode(
         resBody.token.replace(/^Bearer /, '')
